# fix: performance/memory updates