… rebuild/src/functions.cc
… rebuild/src/public_key.h
… rebuild/src/secret_key.cc
… rebuild/src/secret_key.h
… rebuild/src/signature.cc
… rebuild/src/signature.h
… rebuild/test/unit/PublicKey.test.ts
… rebuild/test/unit/SecretKey.test.ts
… rebuild/test/unit/Signature.test.ts